refactor

+++ src/Endgame.cpp
@@ -7,13 +7,13 @@ namespace {
 
     /// Drive a piece towards the edge of the board,
     /// used in KX vs K and KQ vs KR
-    inline int pushToEdge(Square s) {
-        return 90 - (7 * nSqr(edgeDistance(sFile(s))) / 2
-                   + 7 * nSqr(edgeDistance(sRank(s))) / 2);
+    inline i32 pushToEdge(Square s) {
+        return 90 - (7 * i32(nSqr(edgeDistance(sFile(s)))) / 2
+                   + 7 * i32(nSqr(edgeDistance(sRank(s)))) / 2);
     }
     /// Drive a piece towards the corner of the board,
     /// used in KBN vs K to A1H8 corners
-    inline int pushToCorner(Square s) {
+    inline i32 pushToCorner(Square s) {
         return 420 * std::abs(7 - sFile(s) - sRank(s));
     }
     /// Drive a piece close to another piece
